+++ src/network/projet_quizz/frontend/src/composant/molecules/QuestionsLlmImportPanel/QuestionsLlmImportPanel.tsx
@@ -1,15 +1,8 @@
 import { Card } from "../../atomes/Card";
 import { QuestionsLlmImportOptionsPanel } from "../QuestionsLlmImportOptionsPanel";
 import { QuestionsLlmImportPromptPanel } from "../QuestionsLlmImportPromptPanel";
-import type { LlmImportOption } from "../QuestionsLlmImportOptionsPanel";
 import { QUESTIONS_LLM_IMPORT_PANEL_STYLES } from "./QuestionsLlmImportPanel.styles";
-
-export type LlmImportReponse = { texte: string; correcte: boolean };
-export type LlmImportQuestion = { question: string; commentaire: string; reponses: [LlmImportReponse, LlmImportReponse, LlmImportReponse, LlmImportReponse] };
-export type LlmImportCollectionBlock = { nom: string; questions: LlmImportQuestion[] };
-export type LlmImportPayload = { user_id?: number; collections: LlmImportCollectionBlock[]; questions_sans_collection: LlmImportQuestion[] };
-export type LlmImportWorkflow = { buildPrompt: (options: LlmImportOption[]) => string; importFromJson: (importText: string) => Promise<string> };
-export type QuestionsLlmImportPanelProps = { options: LlmImportOption[]; setOptions: (options: LlmImportOption[]) => void; llmImportWorkflow: LlmImportWorkflow };
+import type { QuestionsLlmImportPanelProps } from "./QuestionsLlmImportPanel.types";
 
 export function QuestionsLlmImportPanel({ options, setOptions, llmImportWorkflow }: QuestionsLlmImportPanelProps) {
   const llmPromptFull = llmImportWorkflow.buildPrompt(options);
