+++ src/network/projet_quizz/frontend/src/composant/molecules/QuestionsLlmImportPanel/QuestionsLlmImportPanel.types.ts
@@ -1,1 +1,23 @@
-export type QuestionsLlmImportPanelTypesPlaceholder = Record<string, never>;
+import type { LlmImportOption } from "../QuestionsLlmImportOptionsPanel";
+
+export type LlmImportReponse = { texte: string; correcte: boolean };
+export type LlmImportQuestion = {
+  question: string;
+  commentaire: string;
+  reponses: [LlmImportReponse, LlmImportReponse, LlmImportReponse, LlmImportReponse];
+};
+export type LlmImportCollectionBlock = { nom: string; questions: LlmImportQuestion[] };
+export type LlmImportPayload = {
+  user_id?: number;
+  collections: LlmImportCollectionBlock[];
+  questions_sans_collection: LlmImportQuestion[];
+};
+export type LlmImportWorkflow = {
+  buildPrompt: (options: LlmImportOption[]) => string;
+  importFromJson: (importText: string) => Promise<string>;
+};
+export type QuestionsLlmImportPanelProps = {
+  options: LlmImportOption[];
+  setOptions: (options: LlmImportOption[]) => void;
+  llmImportWorkflow: LlmImportWorkflow;
+};
