+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsLlmImportCard/QuestionsLlmImportCard.types.ts
@@ -1,1 +1,14 @@
-export type QuestionsLlmImportCardTypesPlaceholder = Record<string, never>;
+import type { CollectionUi, QuizzModuleRow, QuizzQuestionRow } from "../../../types/quizz";
+
+export type QuestionsLlmImportCardProps = {
+  data: {
+    targetCollectionNumeric: number | null;
+    collections: CollectionUi[];
+    allModules: QuizzModuleRow[];
+    importTargetModuleId: number | null;
+    questions: QuizzQuestionRow[];
+  };
+  actions: {
+    onImportSuccess: () => void;
+  };
+};
